+++ const_call_17.go
@@ -1,0 +1,8 @@
+//go:build go_tarantool_call_17
+// +build go_tarantool_call_17
+
+package tarantool
+
+const (
+	CallRequest = Call17Request
+)